Malloc from shuffler thread causes problem!!
It looks like the linked list corruption problem is tied to pthread. The
threads use the same heap space, allocated for the program, and calling
malloc from child thread is not safe.

The list which was getting corrupted a lot was also generated by child
thread and it gets corrupted when the child thread calls malloc... Since
we're working with only 'one' shuffler thread (and 3 mapper threads,) it
doesn't mean that the malloc is corrupting the data generated by main
thread.

So, we've changed the List<Key> handling a bit to utilize heap array.
But this time, generating List from heap array is causing problem. When
we try to push an element from a heap array, the malloc call to make a
linked list node fails!!

Valgrind says "Assertion 'bszB_lo == bszB_hi' mismatch. lo = 80, hi =
130079232" which usually happens when we try to access out of index
array. But we're not making such mistake this time...

Let's investigate this problem at home!! Duh~~~