PUB-SUB model is not appropriate.
In fact, multithreading approach isn't so favorable. We may need to
reconsider the approach.

Since we need each worker to be spawned and wait for request, it's
better to use 'fork' (i.e. multiprocessing) approach than must
multithreading since the threads must be 'finished' before requesting
the signal.

Let me think about this.